fix(enrich): three failure-path defects that turn a transient error into a lasting one (#605, #642, #658)
All three are the same shape: the system handles a transient failure by
silently making a worse, permanent decision instead of degrading gracefully.

#605 — DigestEmbedFailed and DigestEnrichFailed literally aliased the same
bit (0x80) in a fully-allocated uint8, so a transient embed failure was
recorded as a permanent ENRICH failure too, and vice versa (skipFlags() in
internal/plugin/retroactive.go excluded the same bit for both processors).
Widens the digest-flags storage (internal/storage's 0x11 DigestFlagsKey) from
a 1-byte to a 2-byte value with a tolerant decode: a legacy 1-byte record's
0x80 is reinterpreted as BOTH failure bits on read (matching the exact
skip-both behavior the old aliasing already produced), so no already-persisted
engram needs a repair-pass migration and none becomes newly stuck or newly
eligible purely from upgrading. Documented as STO-16.

#642 — settles contradictory prior analysis by reading the actual breaker
call site: internal/plugin/enrich/enrich.go's EnrichService.Enrich wraps
pipeline.Run in circuit.Breaker.Do, which records a failure for ANY non-nil
error, including a caller's context.Canceled/DeadlineExceeded surfaced
through providerTransportError's plain %w wrap. d9bef43 (#643) fixed a
different call site (retroactive.go's systemic-error classification) and left
this one untouched. Adds Breaker.DoIgnoring plus RecordIgnore (releases a
consumed HalfOpen probe slot rather than stranding recovery) so a caller
cancellation no longer counts against the circuit that exists to detect a bad
*provider*.

#658 — phase1 handed the embed call the caller's un-divided context. When an
embed backend hung until that context's own deadline, Embed returned the
right degraded result, but by then the shared context was already expired, so
Run's HopDepth>0 traversal deadline check returned ctx.Err() directly before
phase6 ever scored the FTS candidates phase2 already held — the exact
"activation: context deadline exceeded" from the issue. Reserves a
configurable FRACTION (not a fixed duration) of the caller's remaining
deadline for the embed sub-call, so the rest of the pipeline always has
budget left to complete the BM25-only fallback.

#639 deliberately out of scope: PR #644 already fences the retroactive
processor against a concurrent vault clear and is mid-review with the
maintainer, not stale or inadequate.

+++ internal/engine/engine_replay.go
@@ -42,7 +42,7 @@ type EnrichmentCandidate struct {
 	CreatedAt     time.Time
 	UpdatedAt     time.Time
 	MissingStages []string
-	DigestFlags   uint8
+	DigestFlags   uint16
 }
 
 // EnrichmentApplyEntity is one externally generated entity result.
@@ -78,11 +78,11 @@ type EnrichmentApplyResult struct {
 	ID            storage.ULID
 	AppliedStages []string
 	UpdatedAt     time.Time
-	DigestFlags   uint8
+	DigestFlags   uint16
 }
 
 // stageToFlag maps a stage name to its DigestFlag bit.
-var stageToFlag = map[string]uint8{
+var stageToFlag = map[string]uint16{
 	"entities":       plugin.DigestEntities,
 	"relationships":  plugin.DigestRelationships,
 	"classification": plugin.DigestClassified,
@@ -600,11 +600,11 @@ func countNonNilEngrams(engrams []*storage.Engram) int {
 	return n
 }
 
-func normalizeEnrichmentStages(stages []string) (uint8, []string, map[string]bool, error) {
+func normalizeEnrichmentStages(stages []string) (uint16, []string, map[string]bool, error) {
 	if len(stages) == 0 {
 		stages = defaultReplayStages
 	}
-	stageMask := uint8(0)
+	stageMask := uint16(0)
 	validStages := make([]string, 0, len(stages))
 	seen := make(map[string]bool, len(stages))
 	for _, s := range stages {
@@ -631,7 +631,7 @@ func normalizeExplicitEnrichmentStages(stages []string) (map[string]bool, error)
 	return seen, nil
 }
 
-func missingStagesForFlags(flags uint8, stages []string) []string {
+func missingStagesForFlags(flags uint16, stages []string) []string {
 	missing := make([]string, 0, len(stages))
 	for _, stage := range stages {
 		if flags&stageToFlag[stage] == 0 {
